Move internal command parsing helpers to application_commands
